refactor: Orchestrator - Methods of LaunchedContainer objects no longer require passing a launcher instance
`LaunchedContainer` objects now receive a launcher instance during initialization and keep it to use for active methods like `get_log`.
This change was needed to make it easier to create meta-launchers. Previously, a LaunchedContainer derivative objects might not work well when a meta-launcher instance was passed to it instead of a normal launcher they are used to.
This change gives more control to the launchers.