feat: API Server - Added a default dummy implementation of user authentication and authorization
Added a default dummy implementation of the `user_details_getter` for use with `api_router.setup_routes`
This is just a placeholder for user authentication and authorization so that every request has a user name and permissions.
The placeholder function authenticates the user as user with name "admin" and read/write/admin permissions.
In a real multi-user deployment, the `get_user_details` function MUST be replaced with real authentication/authorization based on OAuth or another auth system.