refactor: Orchestrator - Switched to scan the cached execution candidates from newest to oldest
The results are still the same.
This change will become relevant in the future when executions can disable cache reuse. Then, there could be many executions with same cache key. Most older executions will be purged, so we'll have to focus on the group of newer non-purged executions. We'll be able to stop scanning the executions early.