refactor: API-Server - API Router - Split the setup_routes method to prepare for multi-tenancy
The `api_router.setup_routes` method had some aspects that are not compatible with multi-tenancy. For example, it took `db_engine` which is not compatible with each tenant having a different database and `db_engine`.
The `create_db_and_tables` nested function was being called only once during the app startup which is not compatible with multiple per-tenant databases.

Some work remains to convert the launcher parameter for multi-tenancy.